feat: Add checkAdmin middleware for inventory routes
Adding the `checkAdmin` middleware in `invController.js` to validate the account type before accessing certain routes. This middleware checks if the user is logged in and has an account with either "Admin" or "Employee" type, otherwise redirects them to the appropriate page.

Also, modified `server.js` to include the `checkAdmin` middleware in the inventory routes, securing the access to those routes only to authorized users.