feat(controller): Add updateAccount and updateAccountInfo functions
This commit adds two new functions, `updateAccount` and `updateAccountInfo`, to the `accountController.js` file. The `updateAccount` function renders the `edit` view for updating user account information. The `updateAccountInfo` function handles the updating of account information in the database based on user input. It also generates a new JWT token and sets it as a cookie for the user. This commit also updates the exports of the `accountController.js` file to include the new functions.

Refactor the accountRoute.js file

This commit makes changes to the `accountRoute.js` file. It adds two new routes, `/update/:accountId` and `/update/accountInfo`, which correspond to the `updateAccount` and `updateAccountInfo` functions in the `accountController.js` file. These routes handle the rendering of the `edit` view and the submission of updated account information, respectively.

Update the account.ejs view

This commit modifies the `account.ejs` view. It removes the to-do comment and adds a link to the `edit` view with the account ID dynamically inserted.

Create the edit.ejs view

This commit adds a new `edit.ejs` view. This view includes two forms for updating account information and account password. The forms have input fields for the first name, last name, and email, and password, respectively. Additionally, the view displays error messages if there are any validation errors.