feat: Add functionality to update account password
This commit adds the `updateAccountPassword` function to the `accountController.js` file. This function handles the logic for updating the account password by hashing the new password with bcrypt. Any errors that occur during the process are logged to the console and an appropriate error message is sent to the client. This functionality is then integrated into the `accountRoute.js` file by defining the corresponding route `/update/accountPassword`.