Refactor account password update
In `accountController.js`, refactor the `updateAccountPassword` function to improve password update functionality. The function now retrieves the `account_id` from `res.locals.accountData` and sets it to `account_id` variable. It also retrieves `nav` using the `utilities.getNav` function.

The function now performs the actual password update by executing an `UPDATE` query on the `account` table, setting the `account_password` to the updated hashed password. Before hashing, it sets the updated password in `res.locals.accountData.account_password`.

If successful, it logs the updated password with the associated account ID, sets a success flash message notifying the user of successful password update, and renders the "account/account" view with appropriate variables.

In case of error, it logs the specific error message and sets an error flash message with the error details. It then sends a 500 status and displays an error message to the user.